refactor: Move validation to schema
